Moved connector tests to it test package.
