+++ it/connectors/EnrolmentStoreConnectorSpec.scala
@@ -16,17 +16,17 @@
 
 package connectors
 
-import base.{AppWithDefaultMockFixtures, SpecBase}
 import com.github.tomakehurst.wiremock.client.WireMock.{aResponse, get, okJson, urlEqualTo}
-import helper.WireMockServerHandler
+import itbase.ItSpecBase
 import org.scalacheck.Gen
+import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.libs.json.Json
 import play.api.test.Helpers._
 
 import scala.concurrent.Future
 
-class EnrolmentStoreConnectorSpec extends SpecBase with AppWithDefaultMockFixtures with WireMockServerHandler {
+class EnrolmentStoreConnectorSpec extends ItSpecBase with ScalaCheckPropertyChecks {
 
   override def guiceApplicationBuilder(): GuiceApplicationBuilder =
     super
